requirements.txt
annotated-types==0.7.0
asttokens==3.0.0
blinker==1.8.2
Bootstrap-Flask==2.3.3
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.2.0
Flask==2.3.2
Flask-CKEditor==0.5.1
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.56.0
garminconnect==0.2.25
garth==0.5.2
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.67.0rc1
gspread==6.1.4
gunicorn==21.2.0
httplib2==0.22.0
idna==3.10
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
kiwisolver==1.4.8
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.26.0
numpy==2.0.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.2.2
plotly==6.0.0
pluggy==1.5.0
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pyscript==0.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.25
stack-data==0.6.3
toml==0.10.2
traitlets==5.14.3
typer==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.0.0
withings-sync==4.2.5
WTForms==3.0.1
